source/zombie-go/pkg/websocket/oslookup_test.go
package websocket_test

import (
	"github.com/yngvark/gridwalls3/source/zombie-go/pkg/log2"
	"github.com/yngvark/gridwalls3/source/zombie-go/pkg/websocket/http"
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestOslookup(t *testing.T) {

	t.Run("Should parse cors worigins", func(t *testing.T) {
		lg, err := log2.New()
		assert.Nil(t, err)

		m := http.New(lg)
		allowed, err := m.GetAllowedCorsOrigins(osLookupEnv, "TEST_ENV")
		assert.Nil(t, err)

		expected := make(map[string]bool)
		expected["http://localhost:3000"] = true
		expected["https://localhost:3001"] = true

		assert.Equal(t, expected, allowed)
	})
}

func osLookupEnv(_ string) (string, bool) {
	return "http://localhost:3000,https://localhost:3001", true
}
